requirements.txt
Django==2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-modeltranslation==0.17.1
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
